+++ src/OrgStat/CLI.hs
@@ -12,11 +12,11 @@ import Options.Applicative.Simple (Parser, help, long, metavar, strOption, switc
 -- | Read-only arguments that inner application needs (in contrast to,
 -- say, logging severity).
 data CommonArgs = CommonArgs
-    { xdgOpen      :: !Bool
+    { caXdgOpen   :: !Bool
       -- ^ Open report types using xdg-open
-    , selectOutput :: !(Maybe Text)
+    , caOutputs   :: ![Text]
       -- ^ Single output can be selected instead of running all of them.
-    , outputDir    :: !(Maybe FilePath)
+    , caOutputDir :: !(Maybe FilePath)
       -- ^ Output directory for all ... outputs.
     } deriving Show
 
@@ -24,11 +24,12 @@ parseCommonArgs :: Parser CommonArgs
 parseCommonArgs =
     CommonArgs <$>
     switch (long "xdg-open" <> help "Open each report using xdg-open") <*>
-    optional (
+    many (
         fromString <$>
-        strOption (long "select-output" <>
-                   help ("Output name you want to process " <>
-                         "(by default all outputs from conf are processed"))) <*>
+        strOption (long "output" <>
+                   long "select-output" <>
+                   help ("Output name(s) you want to process " <>
+                         "(by default all outputs from conf are selected)"))) <*>
     optional (
         strOption (long "output-dir" <>
                    metavar "FILEPATH" <>