+++ CHANGES.md
@@ -1,3 +1,9 @@
+Unreleased
+==========
+
+* Support selecting multiple outputs in cli: --select-output can be used several times,
+  also it's renamed into --output.
+
 0.1.4
 =====